+++ packages/dso-toolkit/src/components/viewer-grid/viewer-grid.args.ts
@@ -4,6 +4,7 @@ import { ArgTypes } from "@storybook/types";
 import { noControl } from "../../storybook/index.js";
 import { ViewerGridPanelSize, ViewerGridMode, Tab, ViewerGrid } from "./viewer-grid.models.js";
 import { ViewerGridTemplates } from "./viewer-grid.stories-of.js";
+import { AdvancedSelect } from "../advanced-select";
 
 const panelSizes = ["small", "medium", "large"];
 
@@ -32,8 +33,10 @@ export interface ViewerGridArgs {
 export interface ViewerGridDocumentHeaderArgs {
   documentHeaderFeaturesOpen: boolean;
   documentHeaderFeatureAction: HandlerFunction;
-  documentHeaderStatusOpen: boolean;
+  documentHeaderAdvancedSelectOpen: boolean;
   documentHeaderSticky: boolean;
+  documentHeaderAdvancedSelect: AdvancedSelect<unknown>;
+  documentHeaderAdvancedSelectActiveIndex: number;
 }
 
 export const viewerGridArgTypes: ArgTypes<ViewerGridArgs> = {
@@ -134,8 +137,14 @@ export const viewerGridDocumentHeaderArgs: ArgTypes<ViewerGridDocumentHeaderArgs
   documentHeaderFeaturesOpen: {
     type: "boolean",
   },
-  documentHeaderStatusOpen: {
+  documentHeaderAdvancedSelectOpen: {
     type: "boolean",
+  },
+  documentHeaderAdvancedSelectActiveIndex: {
+    name: "Active option",
+    control: {
+      type: "number",
+    },
   },
   documentHeaderFeatureAction: {
     ...noControl,
@@ -143,6 +152,9 @@ export const viewerGridDocumentHeaderArgs: ArgTypes<ViewerGridDocumentHeaderArgs
   },
   documentHeaderSticky: {
     type: "boolean",
+  },
+  documentHeaderAdvancedSelect: {
+    ...noControl,
   },
 };